moved batch api results to check_batch_gen.py
